[refactor] rename header component to v-header to comply with vue style guide
